+++ tests/test_doit_tasks/test_doc.py
@@ -59,7 +59,7 @@ def test_task_cl_bump_pre():
     result = task_cl_bump_pre()
 
     actions = result['actions']
-    assert len(actions) == 4
+    assert len(actions) == 5
     assert isinstance(actions[1][0], type(_move_cl))
     assert 'poetry run cz bump --prerelease' in str(actions[2])
     assert actions[3] == 'git push origin --tags --no-verify'