+++ calcipy/dev/noxfile.py
@@ -32,15 +32,16 @@ with open(path_stdout, 'w') as out:
 
 """
 
+import json
 import re
 import shlex
 from contextlib import suppress
 from pathlib import Path
-from typing import Callable, Dict, List
 from urllib.parse import urlparse
 from urllib.request import url2pathname
 
 from beartype import beartype
+from beartype.typing import Callable, Dict, List
 from loguru import logger
 
 from ..doit_tasks.doit_globals import DG, DoitAction, DoitTask
@@ -222,10 +223,11 @@ if _HAS_TEST_IMPORTS:  # pragma: no cover  # noqa: C901
         """
         session.install('safety', '--upgrade')
         path_report = Path('insecure_report.json').resolve()
-        logger.info(f'Creating safety report: {path_report}')
-        session.run(*shlex.split(f'safety check --full-report --cache --output {path_report} --json'), stdout=True)
-        if path_report.read_text().strip() != '[]':
-            raise RuntimeError(f'Found safety warnings in {path_report}')
+        logger.info('Creating safety report: {path_report}', path_report=path_report)
+        session.run(*shlex.split(f'safety check --full-report --cache --output {path_report} --json'), stdout=False)
+        report_dict = json.loads(path_report.read_text())
+        if matched_vulns := [item for item in report_dict if (item.get('vulns', []) != [])]:
+            raise RuntimeError(f'Found safety warnings in {path_report} for {matched_vulns}')
         path_report.unlink()
 
     @nox_session(python=DG.test.pythons[-1:], reuse_venv=True)