+++ calcipy/file_helpers.py
@@ -5,10 +5,10 @@ import shutil
 import string
 import time
 from pathlib import Path
-from typing import Any, List, Optional
 
 import yaml
 from beartype import beartype
+from beartype.typing import Any, List, Optional
 from loguru import logger
 
 # ----------------------------------------------------------------------------------------------------------------------
@@ -114,9 +114,7 @@ def read_lines(path_file: Path) -> List[str]:
         List[str]: lines of text as list
 
     """
-    if path_file.is_file():
-        return path_file.read_text().split('\n')
-    return []
+    return path_file.read_text().split('\n') if path_file.is_file() else []
 
 
 @beartype